# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  username: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: User;
}

export type BotStatusProperty = typeof BotStatusProperty[keyof typeof BotStatusProperty];


export const BotStatusProperty = {
  inactive: 'inactive',
  connecting: 'connecting',
  connected: 'connected',
  disconnected: 'disconnected',
} as const;

export type SubscriptionPlan = typeof SubscriptionPlan[keyof typeof SubscriptionPlan];


export const SubscriptionPlan = {
  free: 'free',
  basic: 'basic',
  premium: 'premium',
} as const;

export interface Subscription {
  id: string;
  userId: string;
  botId: string;
  plan: SubscriptionPlan;
  startDate: string;
  endDate: string;
  isActive: boolean;
  features: string[];
}

export interface Bot {
  id: string;
  ownerId: string;
  name: string;
  /** @nullable */
  phoneNumber?: string | null;
  status: BotStatusProperty;
  prefix: string;
  createdAt: string;
  subscription?: Subscription;
}

export interface BotInput {
  /** @minLength 1 */
  name: string;
  phoneNumber?: string;
  prefix?: string;
}

export type BotStatusStatus = typeof BotStatusStatus[keyof typeof BotStatusStatus];


export const BotStatusStatus = {
  inactive: 'inactive',
  connecting: 'connecting',
  connected: 'connected',
  disconnected: 'disconnected',
} as const;

export interface BotStatus {
  id: string;
  status: BotStatusStatus;
  /** @nullable */
  phoneNumber?: string | null;
}

export interface BotStats {
  total: number;
  connected: number;
  connecting: number;
  inactive: number;
  disconnected: number;
}

export interface PairingInput {
  phoneNumber: string;
}

export interface PairingResponse {
  code: string;
}

export interface Plan {
  id: string;
  name: string;
  features: string[];
  durationDays: number;
  price: number;
}

export type SubscriptionActivateInputPlan = typeof SubscriptionActivateInputPlan[keyof typeof SubscriptionActivateInputPlan];


export const SubscriptionActivateInputPlan = {
  free: 'free',
  basic: 'basic',
  premium: 'premium',
} as const;

export interface SubscriptionActivateInput {
  botId: string;
  plan: SubscriptionActivateInputPlan;
  durationDays?: number;
}

export interface SubscriptionExtendInput {
  subscriptionId: string;
  durationDays: number;
}