# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Bot Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: bots
    description: Bot management
  - name: subscriptions
    description: Subscription management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List user's bots
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: createBot
      tags: [bots]
      summary: Create a new bot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get bot by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete bot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found

  /bots/{id}/start:
    post:
      operationId: startBot
      tags: [bots]
      summary: Start bot (initiate WhatsApp connection)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot starting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

  /bots/{id}/stop:
    post:
      operationId: stopBot
      tags: [bots]
      summary: Stop bot (disconnect WhatsApp)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

  /bots/{id}/status:
    get:
      operationId: getBotStatus
      tags: [bots]
      summary: Get bot connection status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"

  /bots/{id}/pairing:
    post:
      operationId: requestPairing
      tags: [bots]
      summary: Request pairing code
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairingInput"
      responses:
        "200":
          description: Pairing code requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingResponse"

  /bots/stats:
    get:
      operationId: getBotStats
      tags: [bots]
      summary: Get bot statistics summary for dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bot stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStats"

  /subscriptions/plans:
    get:
      operationId: listPlans
      tags: [subscriptions]
      summary: List available subscription plans
      responses:
        "200":
          description: Available plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"

  /subscriptions/activate:
    post:
      operationId: activateSubscription
      tags: [subscriptions]
      summary: Activate a subscription plan for a bot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionActivateInput"
      responses:
        "200":
          description: Subscription activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  /subscriptions/status/{botId}:
    get:
      operationId: getSubscriptionStatus
      tags: [subscriptions]
      summary: Get subscription status for a bot
      security:
        - bearerAuth: []
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  /subscriptions/extend:
    post:
      operationId: extendSubscription
      tags: [subscriptions]
      summary: Extend subscription (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionExtendInput"
      responses:
        "200":
          description: Extended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string

    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    Bot:
      type: object
      required: [id, ownerId, name, status, prefix, createdAt]
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        phoneNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [inactive, connecting, connected, disconnected]
        prefix:
          type: string
        createdAt:
          type: string
        subscription:
          $ref: "#/components/schemas/Subscription"

    BotInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        phoneNumber:
          type: string
        prefix:
          type: string

    BotStatus:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [inactive, connecting, connected, disconnected]
        phoneNumber:
          type: ["string", "null"]

    BotStats:
      type: object
      required: [total, connected, connecting, inactive, disconnected]
      properties:
        total:
          type: integer
        connected:
          type: integer
        connecting:
          type: integer
        inactive:
          type: integer
        disconnected:
          type: integer

    PairingInput:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string

    PairingResponse:
      type: object
      required: [code]
      properties:
        code:
          type: string

    Plan:
      type: object
      required: [id, name, features, durationDays, price]
      properties:
        id:
          type: string
        name:
          type: string
        features:
          type: array
          items:
            type: string
        durationDays:
          type: integer
        price:
          type: number

    Subscription:
      type: object
      required: [id, userId, botId, plan, startDate, endDate, isActive, features]
      properties:
        id:
          type: string
        userId:
          type: string
        botId:
          type: string
        plan:
          type: string
          enum: [free, basic, premium]
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        features:
          type: array
          items:
            type: string

    SubscriptionActivateInput:
      type: object
      required: [botId, plan]
      properties:
        botId:
          type: string
        plan:
          type: string
          enum: [free, basic, premium]
        durationDays:
          type: integer

    SubscriptionExtendInput:
      type: object
      required: [subscriptionId, durationDays]
      properties:
        subscriptionId:
          type: string
        durationDays:
          type: integer