# lib/api-zod/src/generated/types/bot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BotStatusProperty } from './botStatusProperty';
import type { Subscription } from './subscription';

export interface Bot {
  id: string;
  ownerId: string;
  name: string;
  /** @nullable */
  phoneNumber?: string | null;
  status: BotStatusProperty;
  prefix: string;
  createdAt: string;
  subscription?: Subscription;
}