# lib/api-zod/src/generated/types/botInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BotInput {
  /** @minLength 1 */
  name: string;
  phoneNumber?: string;
  prefix?: string;
}