# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BotStats {
  total: number;
  connected: number;
  connecting: number;
  inactive: number;
  disconnected: number;
}