# lib/api-zod/src/generated/types/subscription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionPlan } from './subscriptionPlan';

export interface Subscription {
  id: string;
  userId: string;
  botId: string;
  plan: SubscriptionPlan;
  startDate: string;
  endDate: string;
  isActive: boolean;
  features: string[];
}