# lib/api-zod/src/generated/types/subscriptionActivateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionActivateInputPlan } from './subscriptionActivateInputPlan';

export interface SubscriptionActivateInput {
  botId: string;
  plan: SubscriptionActivateInputPlan;
  durationDays?: number;
}